pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Yeray Diaz", email = "yeray@softay.es"},
]
dependencies = [
    "fastapi>=0.85.0",
    "sqlalchemy>=1.4.41",
    "psycopg2-binary>=2.9.3",
    "uvicorn>=0.18.3",
    "alembic>=1.8.1",
    "pydantic>=1.10.2",
    "typer>=0.6.1",
    "python-dotenv>=0.21.0",
    "python-dateutil>=2.8.2",
    "pytest>=7.1.3"
]
requires-python = ">=3.10"
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]

[tool.pdm.scripts]
server = "uvicorn src.infrastructure.server.main:app --reload --host 0.0.0.0"
cli = "python -m src.infrastructure.commands.main"
init_alembic = "alembic init alembic"
make_migration = "alembic revision --autogenerate -m"
migrate = "alembic upgrade head"